requirements.txt
ailment==9.0.4663
angr==9.0.4663
appdirs==1.4.4
archinfo==9.0.4663
attrs==19.3.0
Automat==0.8.0
bitstring==3.1.7
blinker==1.4
cachetools==4.1.1
capstone==4.0.2
certifi==2019.11.28
cffi==1.14.3
chardet==3.0.4
claripy==9.0.4663
cle==9.0.4663
Click==7.0
cloud-init==20.2
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
CppHeaderParser==2.7.4
cryptography==2.8
dbus-python==1.2.16
decorator==4.4.2
distro==1.4.0
distro-info===0.23ubuntu1
dpkt==1.9.4
entrypoints==0.3
future==0.18.2
gitdb==4.0.5
GitPython==3.1.11
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
itanium-demangler==1.0
Jinja2==2.10.1
jsonpatch==1.22
jsonpickle==1.4.1
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
llvmcpy==0.1.5
llvmlite==0.34.0
MarkupSafe==1.1.0
more-itertools==4.2.0
mulpyplexer==0.8
netifaces==0.10.4
networkx==2.5
oauthlib==3.1.0
packaging==20.4
parmap==1.5.2
pefile==2019.4.18
plumbum==1.6.9
ply==3.11
progressbar2==3.53.1
protobuf==3.13.0
psutil==5.7.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pyelftools==0.27
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
PySMT==0.9.0
python-apt==2.0.0+ubuntu0.20.4.1
python-debian===0.1.36ubuntu1
python-magic==0.4.18
python-utils==2.4.0
pyvex==9.0.4663
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
rpyc==4.1.5
scandir==1.10.0
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.16.0
simuvex==7.7.9.8
six==1.14.0
smmap==3.0.4
snakeviz==2.1.0
sortedcontainers==2.3.0
ssh-import-id==5.10
systemd-python==234
tornado==6.1
tqdm==4.51.0
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
unicorn==1.0.2rc4
urllib3==1.25.8
wadllib==1.3.3
z3-solver==4.8.9.0
zipp==1.0.0
zope.interface==4.7.1
